refactor: 将主键类型从 Long 更改为 Integer，优化数据模型和数据库访问层

+++ avalon-application/avalon-admin-server/src/main/java/io/github/lishangbu/avalon/admin/service/dataset/TypeService.java
@@ -42,7 +42,7 @@ public interface TypeService {
    *
    * @param id 要删除的 Type 主键
    */
-  void removeById(Long id);
+  void removeById(Integer id);
 
   /**
    * 更新属性类型
